backend/billirae_backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.12"}
psycopg = {extras = ["binary"], version = "^3.2.7"}
motor = "^3.7.0"
pymongo = "^4.12.1"
pydantic = {extras = ["email"], version = "^2.11.4"}
python-jose = {extras = ["cryptography"], version = "^3.4.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.20"
reportlab = "^4.4.0"
python-dotenv = "^1.1.0"
openai = "^1.76.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
